requirements/dev.txt
-r common.txt

django-debug-toolbar==1.9.1
django-extensions==2.0.6
